+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## 2.2.1 - Unreleased
 
 - Add support for `QuerySet.explain()` options.
+- Remove `psycopg2` from `setup.py`'s `install_requires` so that users can
+  install either `psycopg2` or `psycopg2-binary`.
 
 ## 2.2 - 2020-01-15